refactor(digest): 文書分割モジュールを sectioning.py へ改名し Task 2 の blocks.py と衝突させない
